+++ lib/src/data/drivers/firebase/firebase_analytics_driver.dart
@@ -45,7 +45,7 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
         callOptions: AnalyticsCallOptions(global: true),
         parameters: _convertToMapStringObject(event.parameters ?? {}),
       );
-      debugPrint('>>> Analytics Event: ${event.name} <<<');
+      debugPrint('>>> FirebaseAnalyticsDriver.createEvent: ${event.name} <<<');
       return Right(unit);
     } catch (exception, stackTrace) {
       debugPrint('FirebaseAnalyticsDriver.createEvent: $exception');
@@ -87,6 +87,119 @@ class FirebaseAnalyticsDriver extends IFirebaseAnalyticsDriver {
       return Right(unit);
     } catch (exception, stackTrace) {
       debugPrint('FirebaseAnalyticsDriver.setUserProperty: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  AnalyticsEventItem _analyticsEventItem(Map<String, dynamic> i) {
+    return AnalyticsEventItem(
+      parameters: i['parameters'],
+      coupon: i['coupon']?.toString(),
+      itemId: i['itemId']?.toString(),
+      currency: i['currency']?.toString(),
+      itemName: i['itemName']?.toString(),
+      itemBrand: i['itemBrand']?.toString(),
+      itemListId: i['itemListId']?.toString(),
+      locationId: i['locationId']?.toString(),
+      affiliation: i['affiliation']?.toString(),
+      index: int.tryParse(i['index'].toString()),
+      price: num.tryParse(i['price'].toString()),
+      promotionId: i['promotionId']?.toString(),
+      itemVariant: i['itemVariant']?.toString(),
+      creativeName: i['creativeName']?.toString(),
+      creativeSlot: i['creativeSlot']?.toString(),
+      itemCategory: i['itemCategory']?.toString(),
+      itemListName: i['itemListName']?.toString(),
+      itemCategory2: i['itemCategory2']?.toString(),
+      itemCategory3: i['itemCategory3']?.toString(),
+      itemCategory4: i['itemCategory4']?.toString(),
+      itemCategory5: i['itemCategory5']?.toString(),
+      promotionName: i['promotionName']?.toString(),
+      discount: num.tryParse(i['discount'].toString()),
+      quantity: int.tryParse(i['quantity'].toString()),
+    );
+  }
+
+  @override
+  Future<Either<Exception, Unit>> addToCart({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+          .map(_analyticsEventItem)
+          .toList();
+      await instance.logAddToCart(
+        items: items,
+        parameters: params['parameters'],
+        currency: params['currency']?.toString() ?? 'BRL',
+        value: double.tryParse(params['value'].toString()),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.addToCart');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.addToCart: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> beginCheckout({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+          .map(_analyticsEventItem)
+          .toList();
+      await instance.logBeginCheckout(
+        items: items,
+        parameters: params['parameters'],
+        coupon: params['coupon']?.toString(),
+        currency: params['currency']?.toString() ?? 'BRL',
+        value: double.tryParse(params['value'].toString()),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.beginCheckout');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.beginCheckout: $exception');
+      crashLog.capture(exception: exception, stackTrace: stackTrace);
+      return Left(Exception(exception));
+    }
+  }
+
+  @override
+  Future<Either<Exception, Unit>> purchase({
+    required Map<String, dynamic> params,
+  }) async {
+    try {
+      final items = ((params['items'] ?? []) as List<Map<String, dynamic>>)
+          .map(_analyticsEventItem)
+          .toList();
+      await instance.logPurchase(
+        items: items,
+        parameters: params['parameters'],
+        coupon: params['coupon']?.toString(),
+        currency: params['currency'] ?? 'BRL',
+        affiliation: params['affiliation']?.toString(),
+        tax: double.tryParse(params['tax'].toString()),
+        value: double.tryParse(params['value'].toString()),
+        transactionId: params['transactionId']?.toString(),
+        shipping: double.tryParse(params['shipping'].toString()),
+        callOptions: AnalyticsCallOptions(
+          global: params['callOptions']?['global'] ?? false,
+        ),
+      );
+      debugPrint('FirebaseAnalyticsDriver.logPurchase');
+      return Right(unit);
+    } catch (exception, stackTrace) {
+      debugPrint('FirebaseAnalyticsDriver.logPurchase: $exception');
       crashLog.capture(exception: exception, stackTrace: stackTrace);
       return Left(Exception(exception));
     }