+++ lib/src/domain/services/firebase/i_firebase_analytics_service.dart
@@ -12,4 +12,13 @@ abstract class IFirebaseAnalyticsService extends IAppTrackingService {
     required String name,
     required String value,
   });
+  Future<Either<Exception, Unit>> addToCart({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> beginCheckout({
+    required Map<String, dynamic> params,
+  });
+  Future<Either<Exception, Unit>> purchase({
+    required Map<String, dynamic> params,
+  });
 }